requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
google-api-core
python-dotenv
nltk
PyPDF2